+++ sdk/identity/azure_identity/src/process/standard.rs
